Added a code for the Seatworks Section
I have finally added some code into the seatworks section of the program. My biggest gripe now is where to put the object of the Formulas class now and how to update an object's properties. I can sense a solution on the tip of my tongue but I'm asking for your advice on how to do it efficiently.